+++ sound-mind-studio/tests/test_main_window.h
@@ -148,4 +148,5 @@ private slots:
     void settingANewProjectResetsPaintModeToOff();
     void paintingWithTheDefaultToolConfigurationActuallyPaintsSomethingVisible();
     void toggleToolConfigurationPanelShowsAndHidesIt();
+    void paintingTargetsTheSelectedLayerNotNecessarilyTheTopmostOne();
 };